Move server cert tests to server_cert_verifier.rs
