+++ rustls/tests/server_cert_verifier.rs
@@ -11,17 +11,24 @@ use common::{
     do_handshake_until_both_error, do_handshake_until_error, make_client_config,
     make_pair_for_arc_configs, make_pair_for_configs, make_server_config,
 };
-use rustls::client::WebPkiServerVerifier;
+use pki_types::UnixTime;
 use rustls::client::danger::{
     HandshakeSignatureValid, ServerCertVerified, ServerCertVerifier, ServerIdentity,
     SignatureVerificationInput,
 };
-use rustls::server::{ClientHello, ResolvesServerCert};
+use rustls::client::{WebPkiServerVerifier, verify_server_cert_signed_by_trust_anchor};
+use rustls::server::{ClientHello, ParsedCertificate, ResolvesServerCert};
 use rustls::sign::CertifiedKey;
 use rustls::{
-    AlertDescription, CertificateError, CertificateType, ClientConfig, DistinguishedName, Error,
-    InvalidMessage, RootCertStore, ServerConfig,
+    AlertDescription, CertificateError, CertificateType, ClientConfig, ClientConnection,
+    DistinguishedName, Error, ExtendedKeyPurpose, InvalidMessage, RootCertStore, ServerConfig,
+    ServerConnection,
 };
+use rustls_test::{
+    certificate_error_expecting_name, make_client_config_with_verifier, server_name,
+    webpki_server_verifier_builder,
+};
+use webpki::anchor_from_trusted_cert;
 use x509_parser::prelude::FromDer;
 use x509_parser::x509::X509Name;
 
@@ -280,6 +287,348 @@ fn client_can_request_certain_trusted_cas() {
     // For cas_unaware clients, all of them should fail except one that happens to
     // have the cert the server sends
     assert_eq!(cas_unaware_error_count, key_types.len() - 1);
+}
+
+#[test]
+fn client_checks_server_certificate_with_given_ip_address() {
+    fn check_server_name(
+        client_config: Arc<ClientConfig>,
+        server_config: Arc<ServerConfig>,
+        name: &'static str,
+    ) -> Result<(), ErrorFromPeer> {
+        let mut client = ClientConnection::new(client_config, server_name(name)).unwrap();
+        let mut server = ServerConnection::new(server_config).unwrap();
+        do_handshake_until_error(&mut client, &mut server)
+    }
+
+    let provider = provider::default_provider();
+    for kt in KeyType::all_for_provider(&provider) {
+        let server_config = Arc::new(make_server_config(*kt, &provider));
+
+        for version_provider in all_versions(&provider) {
+            let client_config = Arc::new(make_client_config(*kt, &version_provider));
+
+            // positive ipv4 case
+            assert_eq!(
+                check_server_name(client_config.clone(), server_config.clone(), "198.51.100.1"),
+                Ok(()),
+            );
+
+            // negative ipv4 case
+            assert_eq!(
+                check_server_name(client_config.clone(), server_config.clone(), "198.51.100.2"),
+                Err(ErrorFromPeer::Client(Error::InvalidCertificate(
+                    certificate_error_expecting_name("198.51.100.2")
+                )))
+            );
+
+            // positive ipv6 case
+            assert_eq!(
+                check_server_name(client_config.clone(), server_config.clone(), "2001:db8::1"),
+                Ok(()),
+            );
+
+            // negative ipv6 case
+            assert_eq!(
+                check_server_name(client_config.clone(), server_config.clone(), "2001:db8::2"),
+                Err(ErrorFromPeer::Client(Error::InvalidCertificate(
+                    certificate_error_expecting_name("2001:db8::2")
+                )))
+            );
+        }
+    }
+}
+
+#[test]
+fn client_checks_server_certificate_with_given_name() {
+    let provider = provider::default_provider();
+    for kt in KeyType::all_for_provider(&provider) {
+        let server_config = Arc::new(make_server_config(*kt, &provider));
+
+        for version_provider in all_versions(&provider) {
+            let client_config = make_client_config(*kt, &version_provider);
+            let mut client = ClientConnection::new(
+                Arc::new(client_config),
+                server_name("not-the-right-hostname.com"),
+            )
+            .unwrap();
+            let mut server = ServerConnection::new(server_config.clone()).unwrap();
+
+            let err = do_handshake_until_error(&mut client, &mut server);
+            assert_eq!(
+                err,
+                Err(ErrorFromPeer::Client(Error::InvalidCertificate(
+                    certificate_error_expecting_name("not-the-right-hostname.com")
+                )))
+            );
+        }
+    }
+}
+
+#[test]
+fn client_check_server_certificate_ee_revoked() {
+    let provider = provider::default_provider();
+    for kt in KeyType::all_for_provider(&provider) {
+        let server_config = Arc::new(make_server_config(*kt, &provider));
+
+        // Setup a server verifier that will check the EE certificate's revocation status.
+        let crls = vec![kt.end_entity_crl()];
+        let builder = webpki_server_verifier_builder(kt.client_root_store(), &provider)
+            .with_crls(crls)
+            .only_check_end_entity_revocation();
+
+        for version_provider in all_versions(&provider) {
+            let client_config =
+                make_client_config_with_verifier(builder.clone(), &version_provider);
+            let mut client =
+                ClientConnection::new(Arc::new(client_config), server_name("localhost")).unwrap();
+            let mut server = ServerConnection::new(server_config.clone()).unwrap();
+
+            // We expect the handshake to fail since the server's EE certificate is revoked.
+            let err = do_handshake_until_error(&mut client, &mut server);
+            assert_eq!(
+                err,
+                Err(ErrorFromPeer::Client(Error::InvalidCertificate(
+                    CertificateError::Revoked
+                )))
+            );
+        }
+    }
+}
+
+#[test]
+fn client_check_server_certificate_ee_unknown_revocation() {
+    let provider = provider::default_provider();
+    for kt in KeyType::all_for_provider(&provider) {
+        let server_config = Arc::new(make_server_config(*kt, &provider));
+
+        // Setup a server verifier builder that will check the EE certificate's revocation status, but not
+        // allow unknown revocation status (the default). We'll provide CRLs that are not relevant
+        // to the EE cert to ensure its status is unknown.
+        let unrelated_crls = vec![kt.intermediate_crl()];
+        let forbid_unknown_verifier =
+            webpki_server_verifier_builder(kt.client_root_store(), &provider)
+                .with_crls(unrelated_crls.clone())
+                .only_check_end_entity_revocation();
+
+        // Also set up a verifier builder that will allow unknown revocation status.
+        let allow_unknown_verifier =
+            webpki_server_verifier_builder(kt.client_root_store(), &provider)
+                .with_crls(unrelated_crls)
+                .only_check_end_entity_revocation()
+                .allow_unknown_revocation_status();
+
+        for version_provider in all_versions(&provider) {
+            let client_config = make_client_config_with_verifier(
+                forbid_unknown_verifier.clone(),
+                &version_provider,
+            );
+            let mut client =
+                ClientConnection::new(Arc::new(client_config), server_name("localhost")).unwrap();
+            let mut server = ServerConnection::new(server_config.clone()).unwrap();
+
+            // We expect if we use the forbid_unknown_verifier that the handshake will fail since the
+            // server's EE certificate's revocation status is unknown given the CRLs we've provided.
+            let err = do_handshake_until_error(&mut client, &mut server);
+            assert_eq!(
+                err,
+                Err(ErrorFromPeer::Client(Error::InvalidCertificate(
+                    CertificateError::UnknownRevocationStatus
+                )))
+            );
+
+            // We expect if we use the allow_unknown_verifier that the handshake will not fail.
+            let client_config =
+                make_client_config_with_verifier(allow_unknown_verifier.clone(), &version_provider);
+            let mut client =
+                ClientConnection::new(Arc::new(client_config), server_name("localhost")).unwrap();
+            let mut server = ServerConnection::new(server_config.clone()).unwrap();
+            let res = do_handshake_until_error(&mut client, &mut server);
+            assert!(res.is_ok());
+        }
+    }
+}
+
+#[test]
+fn client_check_server_certificate_intermediate_revoked() {
+    let provider = provider::default_provider();
+    for kt in KeyType::all_for_provider(&provider) {
+        let server_config = Arc::new(make_server_config(*kt, &provider));
+
+        // Setup a server verifier builder that will check the full chain revocation status against a CRL
+        // that marks the intermediate certificate as revoked. We allow unknown revocation status
+        // so the EE cert's unknown status doesn't cause an error.
+        let crls = vec![kt.intermediate_crl()];
+        let full_chain_verifier_builder =
+            webpki_server_verifier_builder(kt.client_root_store(), &provider)
+                .with_crls(crls.clone())
+                .allow_unknown_revocation_status();
+
+        // Also set up a verifier builder that will use the same CRL, but only check the EE certificate
+        // revocation status.
+        let ee_verifier_builder = webpki_server_verifier_builder(kt.client_root_store(), &provider)
+            .with_crls(crls.clone())
+            .only_check_end_entity_revocation()
+            .allow_unknown_revocation_status();
+
+        for version_provider in all_versions(&provider) {
+            let client_config = make_client_config_with_verifier(
+                full_chain_verifier_builder.clone(),
+                &version_provider,
+            );
+            let mut client =
+                ClientConnection::new(Arc::new(client_config), server_name("localhost")).unwrap();
+            let mut server = ServerConnection::new(server_config.clone()).unwrap();
+
+            // We expect the handshake to fail when using the full chain verifier since the intermediate's
+            // EE certificate is revoked.
+            let err = do_handshake_until_error(&mut client, &mut server);
+            assert_eq!(
+                err,
+                Err(ErrorFromPeer::Client(Error::InvalidCertificate(
+                    CertificateError::Revoked
+                )))
+            );
+
+            let client_config =
+                make_client_config_with_verifier(ee_verifier_builder.clone(), &version_provider);
+            let mut client =
+                ClientConnection::new(Arc::new(client_config), server_name("localhost")).unwrap();
+            let mut server = ServerConnection::new(server_config.clone()).unwrap();
+            // We expect the handshake to succeed when we use the verifier that only checks the EE certificate
+            // revocation status. The revoked intermediate status should not be checked.
+            let res = do_handshake_until_error(&mut client, &mut server);
+            assert!(res.is_ok())
+        }
+    }
+}
+
+#[test]
+fn client_check_server_certificate_ee_crl_expired() {
+    let provider = provider::default_provider();
+    for kt in KeyType::all_for_provider(&provider) {
+        let server_config = Arc::new(make_server_config(*kt, &provider));
+
+        // Setup a server verifier that will check the EE certificate's revocation status, with CRL expiration enforced.
+        let crls = vec![kt.end_entity_crl_expired()];
+        let enforce_expiration_builder =
+            webpki_server_verifier_builder(kt.client_root_store(), &provider)
+                .with_crls(crls)
+                .only_check_end_entity_revocation()
+                .enforce_revocation_expiration();
+
+        // Also setup a server verifier without CRL expiration enforced.
+        let crls = vec![kt.end_entity_crl_expired()];
+        let ignore_expiration_builder =
+            webpki_server_verifier_builder(kt.client_root_store(), &provider)
+                .with_crls(crls)
+                .only_check_end_entity_revocation();
+
+        for version_provider in all_versions(&provider) {
+            let client_config = make_client_config_with_verifier(
+                enforce_expiration_builder.clone(),
+                &version_provider,
+            );
+            let mut client =
+                ClientConnection::new(Arc::new(client_config), server_name("localhost")).unwrap();
+            let mut server = ServerConnection::new(server_config.clone()).unwrap();
+
+            // We expect the handshake to fail since the CRL is expired.
+            let err = do_handshake_until_error(&mut client, &mut server);
+            assert!(matches!(
+                err,
+                Err(ErrorFromPeer::Client(Error::InvalidCertificate(
+                    CertificateError::ExpiredRevocationListContext { .. }
+                )))
+            ));
+
+            let client_config = make_client_config_with_verifier(
+                ignore_expiration_builder.clone(),
+                &version_provider,
+            );
+            let mut client =
+                ClientConnection::new(Arc::new(client_config), server_name("localhost")).unwrap();
+            let mut server = ServerConnection::new(server_config.clone()).unwrap();
+
+            // We expect the handshake to succeed when CRL expiration is ignored.
+            let res = do_handshake_until_error(&mut client, &mut server);
+            assert!(res.is_ok())
+        }
+    }
+}
+
+/// Simple smoke-test of the webpki verify_server_cert_signed_by_trust_anchor helper API.
+/// This public API is intended to be used by consumers implementing their own verifier and
+/// so isn't used by the other existing verifier tests.
+#[test]
+fn client_check_server_certificate_helper_api() {
+    for kt in KeyType::all_for_provider(&provider::default_provider()) {
+        let chain = kt.chain();
+        let correct_roots = kt.client_root_store();
+        let incorrect_roots = match kt {
+            KeyType::Rsa2048 => KeyType::EcdsaP256,
+            _ => KeyType::Rsa2048,
+        }
+        .client_root_store();
+        // Using the correct trust anchors, we should verify without error.
+        assert!(
+            verify_server_cert_signed_by_trust_anchor(
+                &ParsedCertificate::try_from(chain.first().unwrap()).unwrap(),
+                &correct_roots,
+                &[chain.get(1).unwrap().clone()],
+                UnixTime::now(),
+                webpki::ALL_VERIFICATION_ALGS,
+            )
+            .is_ok()
+        );
+        // Using the wrong trust anchors, we should get the expected error.
+        assert_eq!(
+            verify_server_cert_signed_by_trust_anchor(
+                &ParsedCertificate::try_from(chain.first().unwrap()).unwrap(),
+                &incorrect_roots,
+                &[chain.get(1).unwrap().clone()],
+                UnixTime::now(),
+                webpki::ALL_VERIFICATION_ALGS,
+            )
+            .unwrap_err(),
+            Error::InvalidCertificate(CertificateError::UnknownIssuer)
+        );
+    }
+}
+
+#[test]
+fn client_check_server_valid_purpose() {
+    let chain = KeyType::EcdsaP256.client_chain();
+    let trust_anchor = chain.last().unwrap();
+    let roots = RootCertStore {
+        roots: vec![
+            anchor_from_trusted_cert(trust_anchor)
+                .unwrap()
+                .to_owned(),
+        ],
+    };
+
+    let error = verify_server_cert_signed_by_trust_anchor(
+        &ParsedCertificate::try_from(chain.first().unwrap()).unwrap(),
+        &roots,
+        &[chain.get(1).unwrap().clone()],
+        UnixTime::now(),
+        webpki::ALL_VERIFICATION_ALGS,
+    )
+    .unwrap_err();
+    assert_eq!(
+        error,
+        Error::InvalidCertificate(CertificateError::InvalidPurposeContext {
+            required: ExtendedKeyPurpose::ServerAuth,
+            presented: vec![ExtendedKeyPurpose::ClientAuth],
+        })
+    );
+
+    assert_eq!(
+        format!("{error}"),
+        "invalid peer certificate: certificate does not allow extended key usage for \
+         server authentication, allows client authentication"
+    );
 }
 
 #[derive(Debug, Clone)]
